fix: [WIP] Internal Markdown parser returns HTML directly
